requirements.txt
Flask==3.1.0
google-generativeai==0.8.3
google-ai-generativelanguage==0.6.10
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
protobuf==5.29.2
requests==2.32.3
Werkzeug==3.1.3
Jinja2==3.1.5
itsdangerous==2.2.0
tqdm==4.67.1
grpcio==1.68.1
gunicorn==20.1.0
